Add Playwright support
Add support for [Playwright](https://github.com/microsoft/playwright) pages. The API is very similar to Puppeteer (based on the fact that it has been developed by former Puppeteer engineers). The main difference in the context of `accessibility-checker` is the `evaluate` function:

- Playwright [supports](https://github.com/microsoft/playwright) only a single argument.
- Puppeteer [supports](https://github.com/puppeteer/puppeteer/blob/main/docs/api.md#pageevaluatepagefunction-args) a list of arguments.

In this PR, I've merged multiple arguments into a single object, so the usage is supported by both APIs.

Fixes: #269

Signed-off-by: Darek Kay <hello@darekkay.com>