Fix error and crash due to custom cacheFolder
Code to address the case when a user configures a custom cacheFolder in achecker.yml.

The problem was, the previous code did not account for the fact that the code runs in the accessibility_checker/src/lib folder, so a require( ) has to be relative to this directory, hence the need to go up two directory levels if the cacheFolder is relative path and not a fully qualified path.

This update works for any cacheFolder directory.