Make nodePath Windows-compatible
Remove code that prefixed nodePath with '../../' if nodePath did not begin with '/'. This code worked on a unix-like host, but on Windows it changed an absolute path such as 'c:\Users\...' to '../../c:\Users\..', which was invalid. Removing this code made this module work in tests on both Windows and Mac, provided that config.cacheFolder was set to an absolute path, such as `${__dirname}/temp`. Testing has not been done with relative or default values of config.cacheFolder.