Fix ace_ibma.Checker is not a constructor error
Currently, the generated `ace-node.js` file is updated multiple times in
threads. This commit reads the `ace-node.js` file if it exists and
will create the accessibility checker without overwriting the file.

Another change is to throw error if there was a problem in writing the
`ace-node.js` file instead of attempting to create the constructor from
the non existing file.