Update ruleArchiveVersion to use the commit sha
the github actions set ruleArchiveVersion as <pr_number>/merge, this is
obtained through GITHUB_REF environment variable, and this problems in the github actions.
The version should be a valid file name that can be used in both unix
and non unix filesystems.

This PR sets the commit sha obtained from GITHUB_SHA environment
variable as the ruleArchiveVersion